pyproject.toml
[project]
name = "concepts-builder"
version = "0.1.0"
description = "Extract structured educational content from textbook PDFs using AI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.60.0",
    "pandas>=3.0.0",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.0.0",
    "ruff>=0.14.14",
    "supabase>=2.27.3",
    "uuid6>=2025.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
